+++ src/main/java/com/hiveworkshop/blizzard/casc/io/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * High level APIs for CASC file system interaction.
+ * <p>
+ * These are intended for ease of use, dealing away with many of the low level details required to work with CASC.
+ */
+package com.hiveworkshop.blizzard.casc.io;
