+++ src/main/java/com/hiveworkshop/blizzard/casc/nio/HashMismatchException.java
@@ -1,0 +1,15 @@
+package com.hiveworkshop.blizzard.casc.nio;
+
+import java.io.IOException;
+
+public class HashMismatchException extends IOException {
+	private static final long serialVersionUID = -7133950344327038673L;
+
+	public HashMismatchException() {
+	}
+
+	public HashMismatchException(String message) {
+		super(message);
+	}
+
+}
