--- src/main/java/hiveworkshop/blizzard/casc/io/package-info.java
@@ -1,6 +1,0 @@
-/**
- * High level APIs for CASC file system interaction.
- * <p>
- * These are intended for ease of use, dealing away with many of the low level details required to work with CASC.
- */
-package hiveworkshop.blizzard.casc.io;
